+++ django_grainy_test/urls.py
@@ -6,7 +6,9 @@ from .views import (
     View,
     view,
     Detail,
-    detail
+    detail,
+    DetailExplicit,
+    detail_explicit
 )
 
 router = routers.DefaultRouter()
@@ -16,7 +18,9 @@ router.register(r'a', ModelAViewSet)
 urlpatterns = [
     url(r'^view/', view),
     url(r'^detail/(?P<id>[0-9]+)/$', detail),
+    url(r'^detail_explicit/(?P<id>[0-9]+)/$', detail_explicit),
     url(r'^detail_class/(?P<id>[0-9]+)/$', Detail.as_view()),
+    url(r'^detail_class_explicit/(?P<id>[0-9]+)/$', DetailExplicit.as_view()),
     url(r'^view_class/', View.as_view()),
     url(r'^view_class_json/', JsonView.as_view()),
     url(r'^', include(router.urls)),