+++ examples/grainy_view.py
@@ -35,3 +35,14 @@ class View(BaseView):
 @grainy_view(namespace="detail.{id}")
 def detail_view(request, id):
     return HttpResponse()
+
+# you can also pass through flags for permissions checks
+@grainy_view(
+    namespace="detail.{id}",
+    # require that the user has explicitly set permissions for the namespace
+    explicit=True,
+    # ignore the user's superuser priviledges
+    ignore_grant_all=True
+)
+def detail_view(request, id):
+    return HttpResponse()